Add 64-bit executable build support

+++ src/Audio/Music/musicStateMachine.cpp
@@ -14,7 +14,7 @@ auto MusicStateMachine::getRandomThemeFromState(const std::string& stateName) co
 	auto found = mStates.find(stateName);
 	PH_ASSERT_CRITICAL(found != mStates.end(), "Music state \"" + stateName + "\" does not exist!");
 
-	size_t numberOfThemes = found->second.filepaths.size();
+	int numberOfThemes = static_cast<int>(found->second.filepaths.size());
 	if(numberOfThemes == 1)
 		return std::pair<std::string, float>(found->second.filepaths[0], found->second.volumeMultipliers[0]);
 	else {
